+++ cowsay-log4shell/log4shell/Dockerfile
