+++ cowsay-log4shell/log4shell/src/main/resources/application.properties
