+++ cowsay-log4shell/log4shell/src/main/resources/log4j2.xml
